+++ python/subprojects/common/mlrl/common/learners.py
@@ -9,6 +9,8 @@ from typing import List
 from sklearn.base import BaseEstimator
 from sklearn.utils.validation import check_is_fitted
 
+KWARG_PREDICT_SCORES = 'predict_scores'
+
 
 class NominalAttributeLearner(ABC):
     """
@@ -16,6 +18,111 @@ class NominalAttributeLearner(ABC):
     """
 
     nominal_attribute_indices: List[int] = None
+
+
+class IncrementalLearner(ABC):
+    """
+    A base class for all single- or multi-label classifiers or rankers that support incremental prediction. For example,
+    when dealing with ensemble models that consist of several ensemble members, it is possible to consider only a subset
+    of the ensemble members for prediction.
+    """
+
+    class IncrementalPredictor(ABC):
+        """
+        A base class for all classes that allow to obtain incremental predictions from a `IncrementalLearner`.
+        """
+
+        @abstractmethod
+        def has_next(self) -> bool:
+            """
+            Returns whether there are any remaining ensemble members that have not been used yet or not.
+
+            :return: True, if there are any remaining ensemble members, False otherwise
+            """
+            pass
+
+        @abstractmethod
+        def get_num_next(self) -> int:
+            """
+            Returns the number of remaining ensemble members that have not been used yet.
+
+            :return: The number of remaining ensemble members
+            """
+            pass
+
+        @abstractmethod
+        def apply_next(self, step_size: int):
+            """
+            Updates the current predictions by considering several of the remaining ensemble members. If not enough
+            ensemble members are remaining, only the available ones will be used for updating the current predictions.
+
+            :param step_size:   The number of additional ensemble members to be considered for prediction
+            :return:            A `numpy.ndarray` or `scipy.sparse` matrix of shape `(num_examples, num_labels)`, that
+                                stores the updated prediction for individual examples and labels
+            """
+            pass
+
+    def predict_incrementally(self, x, **kwargs) -> IncrementalPredictor:
+        """
+        Returns an `IncrementalPredictor` that allows to obtain predictions for given query examples incrementally. If
+        the optional keyword argument `predict_scores` is set to `True`, regression scores are obtained instead of
+        binary predictions.
+
+        :param x:   A `numpy.ndarray` or `scipy.sparse` matrix, shape `(num_examples, num_features)`, that stores the
+                    feature values of the query examples
+        :return:    The `IncrementalPredictor` that has been created
+        """
+        check_is_fitted(self)
+
+        if bool(kwargs.get(KWARG_PREDICT_SCORES, False)):
+            return self._predict_scores_incrementally(x, **kwargs)
+        else:
+            return self._predict_labels_incrementally(x, **kwargs)
+
+    def predict_proba_incrementally(self, x, **kwargs) -> IncrementalPredictor:
+        """
+        Returns an `IncrementalPredictor` that allows to obtain probability estimates for given query examples
+        incrementally.
+
+        :param x:   A `numpy.ndarray` or `scipy.sparse` matrix, shape `(num_examples, num_features)`, that stores the
+                    feature values of the query examples
+        :return:    The `IncrementalPredictor` that has been created
+        """
+        check_is_fitted(self)
+        return self._predict_proba_incrementally(x, **kwargs)
+
+    def _predict_labels_incrementally(self, x, **kwargs) -> IncrementalPredictor:
+        """
+        May be overridden by subclasses in order to create an `IncrementalPredictor` that allows to obtain binary
+        predictions for given query examples incrementally.
+
+        :param x:   A `numpy.ndarray` or `scipy.sparse` matrix, shape `(num_examples, num_features)`, that stores the
+                    feature values of the query examples
+        :return:    The `IncrementalPredictor` that has been created
+        """
+        raise RuntimeError('Incremental prediction of binary labels not supported using the current configuration')
+
+    def _predict_scores_incrementally(self, x, **kwargs) -> IncrementalPredictor:
+        """
+        May be overridden by subclasses in order to create an `IncrementalPredictor` that allows to obtain regression
+        scores for given query examples incrementally.
+
+        :param x:   A `numpy.ndarray` or `scipy.sparse` matrix, shape `(num_examples, num_features)`, that stores the
+                    feature values of the query examples
+        :return:    The `IncrementalPredictor` that has been created
+        """
+        raise RuntimeError('Incremental prediction of regression scores not supported using the current configuration')
+
+    def _predict_proba_incrementally(self, x, **kwargs) -> IncrementalPredictor:
+        """
+        May be overridden by subclasses in order to create an `IncrementalPredictor` that allows to obtain probability
+        estimates for given query examples incrementally.
+
+        :param x:   A `numpy.ndarray` or `scipy.sparse` matrix, shape `(num_examples, num_features)`, that stores the
+                    feature values of the query examples
+        :return:    The `IncrementalPredictor` that has been created
+        """
+        raise RuntimeError('Incremental prediction of probabilities not supported using the current configuration')
 
 
 class Learner(BaseEstimator, ABC):
@@ -48,7 +155,7 @@ class Learner(BaseEstimator, ABC):
         """
         check_is_fitted(self)
 
-        if bool(kwargs.get('predict_scores', False)):
+        if bool(kwargs.get(KWARG_PREDICT_SCORES, False)):
             return self._predict_scores(x, **kwargs)
         else:
             return self._predict_labels(x, **kwargs)