+++ python/subprojects/testbed/mlrl/testbed/predictions.py
@@ -15,23 +15,96 @@ from mlrl.testbed.data_splitting import DataSplit, DataType
 from mlrl.testbed.io import SUFFIX_ARFF, get_file_name_per_fold
 
 
+class PredictionScope(ABC):
+    """
+    Provides information about whether predictions have been obtained from a global model or incrementally.
+    """
+
+    @abstractmethod
+    def is_global(self) -> bool:
+        """
+        Returns whether the predictions have been obtained from a global model or not.
+
+        :return: True, if the predictions have been obtained from a global model, False otherwise
+        """
+        pass
+
+    @abstractmethod
+    def get_model_size(self) -> int:
+        """
+        Returns the size of the model from which the prediction have been obtained.
+
+        :return: The size of the model or 0, if the predictions have been obtained from a global model
+        """
+        pass
+
+    @abstractmethod
+    def get_file_name(self, name: str) -> str:
+        """
+        Returns a file name that corresponds to a specific prediction scope.
+
+        :param name:    The name of the file (without suffix)
+        :return:        The file name
+        """
+        pass
+
+
+class GlobalPrediction(PredictionScope):
+    """
+    Provides information about predictions that have been obtained from a global model.
+    """
+
+    def is_global(self) -> bool:
+        return True
+
+    def get_model_size(self) -> int:
+        return 0
+
+    def get_file_name(self, name: str) -> str:
+        return name
+
+
+class IncrementalPrediction(PredictionScope):
+    """
+    Provides information about predictions that have been obtained incrementally.
+    """
+
+    def __init__(self, model_size: int):
+        """
+        :param model_size: The size of the model, the predictions have been obtained from
+        """
+        self.model_size = model_size
+
+    def is_global(self) -> bool:
+        return False
+
+    def get_model_size(self) -> int:
+        return self.model_size
+
+    def get_file_name(self, name: str) -> str:
+        return name + '_model-size-' + str(self.model_size)
+
+
 class PredictionOutput(ABC):
     """
     An abstract base class for all outputs, predictions may be written to.
     """
 
     @abstractmethod
-    def write_predictions(self, meta_data: MetaData, data_split: DataSplit, data_type: DataType, predictions,
-                          ground_truth):
+    def write_predictions(self, meta_data: MetaData, data_split: DataSplit, data_type: DataType,
+                          prediction_scope: PredictionScope, predictions, ground_truth):
         """
         Writes predictions to the output.
 
-        :param meta_data:       The meta-data of the data set
-        :param data_split:      The split of the available data, the predictions and ground truth labels correspond to
-        :param data_type:       Specifies whether the predictions and ground truth labels correspond to the training or
-                                test data
-        :param predictions:     The predictions
-        :param ground_truth:    The ground truth
+        :param meta_data:           The meta-data of the data set
+        :param data_split:          The split of the available data, the predictions and ground truth labels correspond
+                                    to
+        :param data_type:           Specifies whether the predictions and ground truth labels correspond to the training
+                                    or test data
+        :param prediction_scope:    Specifies whether the predictions have been obtained from a global model or
+                                    incrementally
+        :param predictions:         The predictions
+        :param ground_truth:        The ground truth
         """
         pass
 
@@ -41,11 +114,14 @@ class PredictionLogOutput(PredictionOutput):
     Outputs predictions and ground truth labels using the logger.
     """
 
-    def write_predictions(self, meta_data: MetaData, data_split: DataSplit, data_type: DataType, predictions,
-                          ground_truth):
+    def write_predictions(self, meta_data: MetaData, data_split: DataSplit, data_type: DataType,
+                          prediction_scope: PredictionScope, predictions, ground_truth):
         text = 'Ground truth:\n\n' + np.array2string(ground_truth, threshold=sys.maxsize) + '\n\nPredictions:\n\n' \
                + np.array2string(predictions, threshold=sys.maxsize, precision=8, suppress_small=True)
         msg = 'Predictions for ' + data_type.value + ' data'
+
+        if not prediction_scope.is_global():
+            msg += ' using a model of size ' + str(prediction_scope.get_model_size())
 
         if data_split.is_cross_validation_used():
             msg += ' (Fold ' + str(data_split.get_fold() + 1) + ')'
@@ -65,9 +141,10 @@ class PredictionArffOutput(PredictionOutput):
         """
         self.output_dir = output_dir
 
-    def write_predictions(self, meta_data: MetaData, data_split: DataSplit, data_type: DataType, predictions,
-                          ground_truth):
-        file_name = get_file_name_per_fold(data_type.get_file_name('predictions'), SUFFIX_ARFF, data_split.get_fold())
+    def write_predictions(self, meta_data: MetaData, data_split: DataSplit, data_type: DataType,
+                          prediction_scope: PredictionScope, predictions, ground_truth):
+        file_name = get_file_name_per_fold(prediction_scope.get_file_name(data_type.get_file_name('predictions')),
+                                           SUFFIX_ARFF, data_split.get_fold())
         attributes = [Label('Ground Truth ' + label.attribute_name) for label in meta_data.labels]
         labels = [Label('Prediction ' + label.attribute_name) for label in meta_data.labels]
         prediction_meta_data = MetaData(attributes, labels, labels_at_start=False)
@@ -85,16 +162,20 @@ class PredictionPrinter:
         """
         self.outputs = outputs
 
-    def print(self, meta_data: MetaData, data_split: DataSplit, data_type: DataType, predictions, ground_truth):
+    def print(self, meta_data: MetaData, data_split: DataSplit, data_type: DataType, prediction_scope: PredictionScope,
+              predictions, ground_truth):
         """
-        :param meta_data:       The meta-data of the data set
-        :param data_split:      The split of the available data, the predictions and ground truth labels correspond to
-        :param data_type:       Specifies whether the predictions and ground truth labels correspond to the training or
-                                test data
-        :param predictions:     A `numpy.ndarray` or `scipy.sparse` matrix, shape `(num_examples, num_labels)`, that
-                                stores the predictions
-        :param ground_truth:    A `numpy.ndarray` or `scipy.sparse` matrix, shape `(num_examples, num_labels)`, that
-                                stores the ground truth labels
+        :param meta_data:           The meta-data of the data set
+        :param data_split:          The split of the available data, the predictions and ground truth labels correspond
+                                    to
+        :param data_type:           Specifies whether the predictions and ground truth labels correspond to the training
+                                    or test data
+        :param prediction_scope:    Specifies whether the predictions have been obtained from a global model or
+                                    incrementally
+        :param predictions:         A `numpy.ndarray` or `scipy.sparse` matrix, shape `(num_examples, num_labels)`, that
+                                    stores the predictions
+        :param ground_truth:        A `numpy.ndarray` or `scipy.sparse` matrix, shape `(num_examples, num_labels)`, that
+                                    stores the ground truth labels
         """
         for output in self.outputs:
-            output.write_predictions(meta_data, data_split, data_type, predictions, ground_truth)
+            output.write_predictions(meta_data, data_split, data_type, prediction_scope, predictions, ground_truth)